Move TMDBSwift Tests
